requirements.txt
Flask
pdfplumber
docx2txt
spacy
nltk
sentence-transformers
scikit-learn
gunicorn
sentencepiece